# pkg/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Optrispace Server API",
        "title": "Optrispace API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/applications": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Returns applications list for the current user — as an applicant or as a job creator",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "application"
                ],
                "summary": "List applications by current authenticated user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Application"
                            }
                        }
                    },
                    "401": {
                        "description": "user not authorized",
                        "schema": {
                            "$ref": "#/definitions/model.BackendError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/echo.HTTPError"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/applications/my": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Returns applications list made by the current user as an applicant",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "application"
                ],
                "summary": "List applications were applied by the current authenticated user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Application"
                            }
                        }
                    },
                    "401": {
                        "description": "user not authorized",
                        "schema": {
                            "$ref": "#/definitions/model.BackendError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/echo.HTTPError"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/applications/{id}": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Applicant create new application for a job",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "application"
                ],
                "summary": "Get an application",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Application ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Application"
                        }
                    },
                    "401": {
                        "description": "user not authorized",
                        "schema": {
                            "$ref": "#/definitions/model.BackendError"
                        }
                    },
                    "404": {
                        "description": "application not found",
                        "schema": {
                            "$ref": "#/definitions/model.BackendError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/echo.HTTPError"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/applications/{id}/chat": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Performer or customer is getting chat for this application",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "application",
                    "chat"
                ],
                "summary": "Get chat for an application",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Application ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Chat"
                        }
                    },
                    "401": {
                        "description": "user not authorized",
                        "schema": {
                            "$ref": "#/definitions/model.BackendError"
                        }
                    },
                    "404": {
                        "description": "application not found or user has no permissions for view chat",
                        "schema": {
                            "$ref": "#/definitions/model.BackendError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/echo.HTTPError"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/jobs/{job_id}/applications": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Returns applications list for the job by job_id. The current user MUST be an applicant or a job creator.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "application",
                    "job"
                ],
                "summary": "List applications for the job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "job_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Application"
                            }
                        }
                    },
                    "401": {
                        "description": "user not authorized",
                        "schema": {
                            "$ref": "#/definitions/model.BackendError"
                        }
                    },
                    "404": {
                        "description": "job not found",
                        "schema": {
                            "$ref": "#/definitions/model.BackendError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/echo.HTTPError"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Applicant create new application for a job",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "application",
                    "job"
                ],
                "summary": "Make new application for a job",
                "parameters": [
                    {
                        "description": "New application request",
                        "name": "application",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.newApplication"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Job ID to apply",
                        "name": "job_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Application"
                        }
                    },
                    "401": {
                        "description": "user not authorized",
                        "schema": {
                            "$ref": "#/definitions/model.BackendError"
                        }
                    },
                    "404": {
                        "description": "job with specified ID is not found",
                        "schema": {
                            "$ref": "#/definitions/model.BackendError"
                        }
                    },
                    "422": {
                        "description": "validation failed (details in response)",
                        "schema": {
                            "$ref": "#/definitions/model.BackendError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/echo.HTTPError"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.newApplication": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                }
            }
        },
...
        "echo.HTTPError": {
            "type": "object",
            "properties": {
                "message": {}
            }
        },
        "model.Application": {
            "type": "object",
            "properties": {
                "applicant": {
                    "$ref": "#/definitions/model.JobApplicant"
                },
                "comment": {
                    "type": "string"
                },
                "contract": {
                    "$ref": "#/definitions/model.Contract"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "job": {
                    "$ref": "#/definitions/model.Job"
                },
                "price": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.BackendError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "tech_info": {
                    "type": "string"
                }
            }
        },
        "model.Chat": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Message"
                    }
                },
                "topic": {
                    "type": "string"
                }
            }
        },
        "model.Contract": {
            "type": "object",
            "properties": {
                "application": {
                    "$ref": "#/definitions/model.Application"
                },
                "contract_address": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "customer": {
                    "$ref": "#/definitions/model.Person"
                },
                "customer_address": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "performer": {
                    "$ref": "#/definitions/model.Person"
                },
                "performer_address": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "model.Job": {
            "type": "object",
            "properties": {
                "applications_count": {
                    "type": "integer"
                },
                "budget": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "customer": {
                    "$ref": "#/definitions/model.Person"
                },
                "description": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.JobApplicant": {
            "type": "object",
            "properties": {
                "display_name": {
                    "type": "string"
                },
                "ethereum_address": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "resources": {
                    "type": "string"
                }
            }
        },
        "model.Message": {
            "type": "object",
            "properties": {
                "chat_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "model.Person": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "display_name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "ethereum_address": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_admin": {
                    "type": "boolean"
                },
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "realm": {
                    "type": "string"
                },
                "resources": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerToken": {
            "description": "Bearer token in Authorization header",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}